Create CovidDeaths_Exploration.sql
Based on the official Covid data obtained from https://ourworldindata.org/covid-deaths, I conducted an exploratory analysis to examine the details of each country and continent regarding death cases caused by Covid-19. The purpose of this analysis was to identify trends, gain insights, and assess the extent of damage caused by the virus specifically for the period between 2020 and 2021.