CovidVaccination_Exploration.sql
Based on the official Covid data obtained from https://ourworldindata.org/covid-deaths, I conducted an exploratory analysis to examine the details of each country and continent regarding vaccinations in order to contain the spread of Covid-19. The purpose of this analysis was to identify trends, gain insights, and assess the extent of damage caused by the virus specifically within a specific period.