Raw Data
This Excel sheet comprises customer details of house purchases in a specific region of the US. However, the data in this sheet contains errors, discrepancies, and inconsistencies, rendering it as 'dirty' data. As a data analyst, working with such data was not feasible, necessitating the need to perform data cleaning procedures to prepare it for analysis. In a separate file, you will find the cleaned version of the data, which is ready for further analysis.